Split diagnostics and the window into one module per responsibility
src/main/diagnostics.ts held the flight recorder, the span factories, the OS
and process samplers, the Chromium trace, and the zip export behind three
module-level mutables that every one of them read and wrote. Each piece now
has a module under src/main/diagnostics/ and one owner for its state:
recorder.ts holds the single FlightRecorder and the primitive recording API,
capture.ts owns the capture level and the raw trace, spans.ts the closed
begin/end pairs, samplers.ts the machine observations and the environment
document, renderer.ts what the renderer reports, and export.ts the .gwdiag.
The recorder class and the JSONL reader move in beside them as
flight-recorder.ts and report.ts. src/main/diagnostics.ts stays the one entry
point the rest of main imports, and composes start and stop.

src/main/window.ts also carried the whole application menu template and the
problem-report dialogs. window-menu.ts and problem-report.ts take those;
window.ts keeps window lifecycle and state persistence, and the menu asks for
a window-state reset rather than reaching into it. resetGameInput moves to
renderer-commands.ts, which already owns main-to-renderer commands.

No behaviour changes: the public surface, the recorded events, and the export
are the same, and the two policy tests that read these sources follow them.
