+++ tests/unit/diagnostic-report.test.ts
@@ -7,7 +7,7 @@ import { describe, it } from "node:test";
 import {
   buildDiagnosticReport,
   previousAbnormalSession,
-} from "../../src/main/diagnostic-report.ts";
+} from "../../src/main/diagnostics/report.ts";
 import type { DiagnosticSummary } from "../../src/shared/diagnostics.ts";
 
 function event(
